+++ SentinelOne/main.py
@@ -1,14 +1,15 @@
-from sentinelone_module.agents.isolation import DeisolateEndpointAction, IsolateEndpointAction
 from sentinelone_module.agents.init_scan import InitiateScanAction
-from sentinelone_module.iocs.create_iocs import CreateIOCsAction
-from sentinelone_module.threats.create_threat_note import CreateThreatNoteAction
-from sentinelone_module.threats.update_threat_incident import UpdateThreatIncidentAction
+from sentinelone_module.agents.isolation import DeisolateEndpointAction, IsolateEndpointAction
 from sentinelone_module.base import SentinelOneModule
 from sentinelone_module.deep_visibility.consumer import DeepVisibilityTrigger
 from sentinelone_module.deep_visibility.query import QueryDeepVisibilityAction
+from sentinelone_module.iocs.create_iocs import CreateIOCsAction
 from sentinelone_module.logs.connector import SentinelOneLogsConnector
 from sentinelone_module.rso.execute import ExecuteRemoteScriptAction
+from sentinelone_module.singularity.connectors import SingularityIdentityConnector
+from sentinelone_module.threats.create_threat_note import CreateThreatNoteAction
 from sentinelone_module.threats.get_malwares import GetMalwaresAction
+from sentinelone_module.threats.update_threat_incident import UpdateThreatIncidentAction
 
 if __name__ == "__main__":
     module = SentinelOneModule()
@@ -23,4 +24,5 @@ if __name__ == "__main__":
     module.register(UpdateThreatIncidentAction, "sentinelone_threat_update_incident")
     module.register(DeepVisibilityTrigger, "sentinelone_deep_visibility_consumer")
     module.register(SentinelOneLogsConnector, "sentinelone_log_connector")
+    module.register(SingularityIdentityConnector, "sentinelone_identity_connector")
     module.run()