+++ SentinelOne/sentinelone_module/iocs/create_iocs.py
@@ -1,6 +1,7 @@
 from collections import defaultdict
 
-from management.mgmtsdk_v2_1.services.threat_intelligence import IocQueryFilter, Ioc
+import pandas as pd
+from management.mgmtsdk_v2_1.services.threat_intelligence import Ioc, IocQueryFilter
 from pydantic import BaseModel
 
 from sentinelone_module.base import SentinelOneAction